Fix shader rendering
This is a constant source of issues... we really need to add testing around it.

First off, ALL operations should be done on unpremul, since that's the native format of the DDS textures we read in.

Secondly, we should treat all textures/tiles as immutable and only perform mutations on canvases or copied bitmaps.

Shader issues are common and frustrating. We need to check that:
  * Shaders are applied (changing haircolor/eyecolor results in different output)
  * Shader bitmaps are correctly loaded and positioned (scattered lines indicate a bug where the shader logic was run without having the original texture copied first)
  * Shader colors are applied (jet black output indicates shading was done incorrectly -- premul issue?)